extract proxy catchers
